refactor (Response, Request) : Deduplication, optimization operations
