+++ src/main/java/com/levelUpToast/levelUpToast/config/InitData.java
@@ -1,26 +1,23 @@
 package com.levelUpToast.levelUpToast.config;
 
-import com.levelUpToast.levelUpToast.domain.member.Authority;
-import com.levelUpToast.levelUpToast.domain.member.Member;
-import com.levelUpToast.levelUpToast.domain.product.Product;
-import com.levelUpToast.levelUpToast.domain.product.buyoption.BuyOption;
-import com.levelUpToast.levelUpToast.domain.product.fundinginfo.FundingInfo;
-import com.levelUpToast.levelUpToast.domain.product.productinfo.ProductInfo;
-import com.levelUpToast.levelUpToast.domain.product.reviwe.Review;
-import com.levelUpToast.levelUpToast.domain.product.tag.Tag;
+import com.levelUpToast.levelUpToast.domain.model.member.Authority;
+import com.levelUpToast.levelUpToast.domain.model.member.Member;
+import com.levelUpToast.levelUpToast.domain.model.product.Product;
+import com.levelUpToast.levelUpToast.domain.model.product.buyoption.BuyOption;
+import com.levelUpToast.levelUpToast.domain.model.product.fundinginfo.FundingInfo;
+import com.levelUpToast.levelUpToast.domain.model.product.productinfo.ProductInfo;
+import com.levelUpToast.levelUpToast.domain.model.product.reviwe.Review;
+import com.levelUpToast.levelUpToast.domain.model.product.tag.Tag;
 import com.levelUpToast.levelUpToast.domain.repository.mainRepository.MainRepository;
 import com.levelUpToast.levelUpToast.domain.repository.memberRepository.memberRepositoryInf.MemberRepository;
 import com.levelUpToast.levelUpToast.domain.repository.productRepository.productRepositoryInf.ProductRepository;
-import com.levelUpToast.levelUpToast.domain.repository.vendorRepository.vendorRepsitoryInf.VendorRepository;
-import com.levelUpToast.levelUpToast.domain.vendor.Vendor;
-import com.levelUpToast.levelUpToast.service.mainPage.MainService;
+import com.levelUpToast.levelUpToast.domain.model.vendor.Vendor;
 import com.levelUpToast.levelUpToast.service.vendor.vendorServiceInf.VendorService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import javax.annotation.PostConstruct;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Calendar;
 
 @Component
 @RequiredArgsConstructor
