--- src/main/java/com/levelUpToast/levelUpToast/domain/dataForm/mainPage/MainPageResponseForm.java
@@ -1,14 +1,0 @@
-package com.levelUpToast.levelUpToast.domain.dataForm.mainPage;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-import java.util.Map;
-
-@Data
-@AllArgsConstructor
-public class MainPageResponseForm {
-    private int detailCode;
-    private String message;
-    private Map<String, Object> data;
-}
