+++ src/main/java/com/levelUpToast/levelUpToast/domain/dataForm/requestForm/img/ImgRequestForm.java
@@ -1,4 +1,4 @@
-package com.levelUpToast.levelUpToast.domain.dataForm.img;
+package com.levelUpToast.levelUpToast.domain.dataForm.requestForm.img;
 
 import lombok.Data;
 import org.springframework.web.multipart.MultipartFile;
