--- src/main/java/com/levelUpToast/levelUpToast/domain/member/Authority.java
@@ -1,3 +1,0 @@
-package com.levelUpToast.levelUpToast.domain.member;
-
-public enum Authority { MEM , ADMIN }
