+++ src/main/java/com/levelUpToast/levelUpToast/domain/model/member/Authority.java
@@ -1,0 +1,3 @@
+package com.levelUpToast.levelUpToast.domain.model.member;
+
+public enum Authority { MEM , ADMIN }
