+++ src/main/java/com/levelUpToast/levelUpToast/domain/model/product/tag/Tag.java
@@ -1,0 +1,3 @@
+package com.levelUpToast.levelUpToast.domain.model.product.tag;
+
+public enum Tag {VEGETABLE, FRUIT}
