--- src/main/java/com/levelUpToast/levelUpToast/domain/product/tag/Tag.java
@@ -1,3 +1,0 @@
-package com.levelUpToast.levelUpToast.domain.product.tag;
-
-public enum Tag {VEGETABLE, FRUIT}
